Add --no-multi to parse_mrts.py script

+++ dnas/scripts/parse_mrts.py
@@ -236,43 +236,50 @@ def parse_file(
             f"filename is not a string: {type(filename)}"
         )
 
-    no_cpu =  multiprocessing.cpu_count()
-    Pool = multiprocessing.Pool(no_cpu)
     mrt_a = mrt_archives()
-
     logging.info(f"Processing {filename}...")
 
-    splitter = mrt_splitter(filename)
-    num_entries, file_chunks = splitter.split(
-        no_chunks=no_cpu,
-        outdir=cfg.SPLIT_DIR
-    )
-    try:
-        splitter.close()
-    except StopIteration:
-        pass
+    if multi:
+        no_cpu =  multiprocessing.cpu_count()
+        Pool = multiprocessing.Pool(no_cpu)
 
-    if mrt_a.is_rib_from_filename(filename):
-        mrt_chunks = Pool.map(mrt_parser.parse_rib_dump, file_chunks)
-    else:
-        mrt_chunks = Pool.map(mrt_parser.parse_upd_dump, file_chunks)
+        splitter = mrt_splitter(filename)
+        num_entries, file_chunks = splitter.split(
+            no_chunks=no_cpu,
+            outdir=cfg.SPLIT_DIR
+        )
+        try:
+            splitter.close()
+        except StopIteration:
+            pass
 
-    Pool.close()
+        if mrt_a.is_rib_from_filename(filename):
+            mrt_chunks = Pool.map(mrt_parser.parse_rib_dump, file_chunks)
+        else:
+            mrt_chunks = Pool.map(mrt_parser.parse_upd_dump, file_chunks)
 
-    if not keep_chunks:
-        for i in range(0, len(file_chunks)):
-            if cfg.SPLIT_DIR:
-                os.remove(
-                    os.path.join(
-                        cfg.SPLIT_DIR, os.path.basename(file_chunks[i])
+        Pool.close()
+
+        if not keep_chunks:
+            for i in range(0, len(file_chunks)):
+                if cfg.SPLIT_DIR:
+                    os.remove(
+                        os.path.join(
+                            cfg.SPLIT_DIR, os.path.basename(file_chunks[i])
+                        )
                     )
-                )
-            else:
-                os.remove(file_chunks[i])
+                else:
+                    os.remove(file_chunks[i])
 
-    mrt_s = mrt_stats()
-    for chunk in mrt_chunks:
-        mrt_s.add(chunk)
+        mrt_s = mrt_stats()
+        for chunk in mrt_chunks:
+            mrt_s.add(chunk)
+
+    else:
+        if mrt_a.is_rib_from_filename(filename):
+            mrt_s = mrt_parser.parse_rib_dump(filename)
+        else:
+            mrt_s = mrt_parser.parse_upd_dump(filename)
 
     return mrt_s
 
@@ -541,7 +548,7 @@ def main():
             "using --single!"
         )
 
-    args["multi"] = not args["no_mult"]
+    args["multi"] = not args["no_multi"]
 
     if args["continuous"]:
         continuous(args)
